Working average calculator

+++ app/src/main/java/comp3350/student_echo/objects/CourseReview.java
@@ -3,7 +3,7 @@ package comp3350.student_echo.objects;
 public class CourseReview extends Review {
     Course course;
 
-    public CourseReview(Course course, String comment, float overallRating, float difficultyRating) {
+    public CourseReview(Course course, String comment, int overallRating, int difficultyRating) {
         super(comment, overallRating, difficultyRating);
         this.course = course;
     }
